Refactor onboarding route group
