+++ src/app/(auth)/onboarding/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '../components/ui/Button';
+import Button from '@/components/ui/Button';
 
 export default function OnboardingPage() {
   return (
