+++ src/app/(main)/cognitive-mirror/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Card from '../../components/ui/Card';
+import Card from '@/components/ui/Card';
 
 export default function CognitiveMirrorPage() {
   return (
