+++ src/app/(main)/narrator/page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Card from '../../components/ui/Card';
-import Button from '../../components/ui/Button';
+import Card from '@/components/ui/Card';
+import Button from '@/components/ui/Button';
 
 export default function NarratorPage() {
   return (
