+++ src/app/(main)/profile-and-privacy/page.tsx
