+++ src/app/(main)/record/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import Button from '../../components/ui/Button';
+import Button from '@/components/ui/Button';
 
 export default function RecordPage() {
   const [isRecording, setIsRecording] = React.useState(false);
